refactor(struct): extract EPUB citations and links tests
